+++ Foundation/NSArchiver.m
@@ -5,8 +5,6 @@ Permission is hereby granted, free of charge, to any person obtaining a copy of 
 The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
 
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
-
-// Original - Christopher Lloyd <cjwl@objc.net>
 #import <Foundation/NSArchiver.h>
 #import <Foundation/NSRaise.h>
 #import <Foundation/NSMutableData.h>
@@ -65,7 +63,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
                format:@"NSArchiver cannot encode type=%s",type];
 }
 
--(void)_ensureLength:(unsigned)length {
+-(void)_ensureLength:(NSUInteger)length {
    [_data setLength:[_data length]+length];
    _bytes=[_data mutableBytes];
 }
@@ -126,7 +124,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
    _bytes[_position++]=value&0xFF;
 }
 
--(void)_appendBytes:(const char *)bytes length:(unsigned)length {
+-(void)_appendBytes:(const char *)bytes length:(NSUInteger)length {
    int i;
 
    if(_pass==0)
@@ -386,7 +384,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 }
 
--(void)encodeBytes:(const void *)byteaddr length:(unsigned)length {
+-(void)encodeBytes:(const void *)byteaddr length:(NSUInteger)length {
    [self _appendWordFour:length];
    [self _appendBytes: byteaddr length:length];
 }
