+++ Foundation/NSArray/NSArray_placeholder.m
@@ -6,14 +6,13 @@ The above copyright notice and this permission notice shall be included in all c
 
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
-// Original - Christopher Lloyd <cjwl@objc.net>
 #import "NSArray_placeholder.h"
 #import "NSArray_concrete.h"
 #import <Foundation/NSRaise.h>
 
 @implementation NSArray_placeholder
 
--initWithObjects:(id *)objects count:(unsigned)count {
+-initWithObjects:(id *)objects count:(NSUInteger)count {
    NSDeallocateObject(self);
    return NSArray_concreteNew(NULL,objects,count);
 }
