+++ Foundation/NSArray/NSEnumerator_arrayReverse.h
@@ -11,7 +11,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 @class NSArray;
 
 @interface NSEnumerator_arrayReverse : NSEnumerator {
-   int      _index;
+   NSInteger _index;
    NSArray *_array;
 }
