+++ Foundation/NSArray/NSMutableArray_concrete.h
@@ -9,12 +9,12 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #import <Foundation/NSArray.h>
 
 @interface NSMutableArray_concrete : NSMutableArray {
-   unsigned _count;
-   unsigned _capacity;
+   NSUInteger _count;
+   NSUInteger _capacity;
    id      *_objects;
 }
 
 @end
 
-NSArray *NSMutableArray_concreteNew(NSZone *zone,id *objects,unsigned count);
-NSArray *NSMutableArray_concreteNewWithCapacity(NSZone *zone,unsigned capacity);
+NSArray *NSMutableArray_concreteNew(NSZone *zone,id *objects,NSUInteger count);
+NSArray *NSMutableArray_concreteNewWithCapacity(NSZone *zone,NSUInteger capacity);
