+++ Foundation/NSAttributedString/NSAttributedString_manyAttributes.m
@@ -6,7 +6,6 @@ The above copyright notice and this permission notice shall be included in all c
 
 THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
-// Original - Christopher Lloyd <cjwl@objc.net>
 #import <Foundation/NSAttributedString_manyAttributes.h>
 #import <Foundation/NSRaise.h>
 #import <Foundation/NSArray.h>
@@ -15,8 +14,8 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 @implementation NSAttributedString_manyAttributes
 
 -initWithAttributedString:(NSAttributedString *)other {
-   unsigned location=0;
-   unsigned limit=[other length];
+   NSUInteger location=0;
+   NSUInteger limit=[other length];
 
    _string=[[other string] copy];
    _rangeToAttributes=NSCreateRangeToCopiedObjectEntries(0);
@@ -43,7 +42,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
    return _string;
 }
 
--(NSDictionary *)attributesAtIndex:(unsigned)location
+-(NSDictionary *)attributesAtIndex:(NSUInteger)location
    effectiveRange:(NSRange *)effectiveRangep {
    NSDictionary *result;
